move get_shifts to module, rename module
