--- lib/encryptor.rb
@@ -1,5 +1,0 @@
-require 'date'
-
-module Encryptor
-
-end
