--- spec/encryptor_spec.rb
@@ -1,13 +1,0 @@
-# require './lib/encryptor'
-#
-# describe Encryptor do
-#   before(:each) do
-#     @encryptor = Encryptor.new
-#   end
-#
-#   it 'exists' do
-#     expect(@encryptor).to be_a(Encryptor)
-#   end
-#
-#
-# end
